Fix remaining quantized_vector references in tests and benchmarks
- Removed quantized_vector field from integration tests
- Removed quantized_vector from benchmark files
- Fixed version type mismatches (i64 to u32)
- Reduced compilation errors from 38 to 4

+++ src/storage/engines/core/formats/columnar/tests.rs
@@ -868,7 +868,7 @@ fn create_test_records(count: usize) -> Vec<VectorRecord> {
                 timestamp: (1000 + i) as i64,
                 updated_at: None,
                 expires_at: None,
-                version: Some(((i % 5) + 1) as i64),
+                version: Some(((i % 5) + 1) as u32),
                 source: None,
             }
         })
